+++ tests/sdk/pix_fraud_log_test.go
@@ -1,0 +1,186 @@
+package sdk
+
+import (
+	"github.com/starkinfra/sdk-go/starkinfra"
+	PixFraudLog "github.com/starkinfra/sdk-go/starkinfra/pixfraud/log"
+	"github.com/starkinfra/sdk-go/tests/utils"
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestPixFraudLogQuery(t *testing.T) {
+
+	starkinfra.User = utils.ExampleProject
+
+	limit := 10
+	var params = map[string]interface{}{}
+	params["limit"] = limit
+
+	logs, errorChannel := PixFraudLog.Query(params, nil)
+loop:
+	for {
+		select {
+		case err := <-errorChannel:
+			if err.Errors != nil {
+				for _, e := range err.Errors {
+					t.Errorf("code: %s, message: %s", e.Code, e.Message)
+				}
+			}
+		case log, ok := <-logs:
+			if !ok {
+				break loop
+			}
+			assert.NotNil(t, log.Id)
+		}
+	}
+}
+
+func TestPixFraudLogPage(t *testing.T) {
+
+	starkinfra.User = utils.ExampleProject
+
+	limit := 10
+	var params = map[string]interface{}{}
+	params["limit"] = limit
+
+	logs, cursor, err := PixFraudLog.Page(params, nil)
+	if err.Errors != nil {
+		for _, e := range err.Errors {
+			t.Errorf("code: %s, message: %s", e.Code, e.Message)
+		}
+	}
+
+	for _, log := range logs {
+		assert.NotNil(t, log.Id)
+	}
+	assert.NotNil(t, cursor)
+}
+
+func TestPixFraudLogGet(t *testing.T) {
+
+	starkinfra.User = utils.ExampleProject
+
+	limit := 10
+	var paramsQuery = map[string]interface{}{}
+	paramsQuery["limit"] = limit
+
+	var logList []PixFraudLog.Log
+
+	logs, errorChannel := PixFraudLog.Query(paramsQuery, nil)
+loop:
+	for {
+		select {
+		case err := <-errorChannel:
+			if err.Errors != nil {
+				for _, e := range err.Errors {
+					t.Errorf("code: %s, message: %s", e.Code, e.Message)
+				}
+			}
+		case log, ok := <-logs:
+			if !ok {
+				break loop
+			}
+			logList = append(logList, log)
+		}
+	}
+
+	for _, log := range logList {
+		getLog, err := PixFraudLog.Get(log.Id, nil)
+		if err.Errors != nil {
+			for _, e := range err.Errors {
+				t.Errorf("code: %s, message: %s", e.Code, e.Message)
+			}
+		}
+		assert.NotNil(t, getLog.Id)
+	}
+
+	assert.Equal(t, limit, len(logList))
+}
+
+func TestPixFraudLogAttributes(t *testing.T) {
+
+	starkinfra.User = utils.ExampleProject
+
+	limit := 10
+	var params = map[string]interface{}{}
+	params["limit"] = limit
+
+	logs, errorChannel := PixFraudLog.Query(params, nil)
+loop:
+	for {
+		select {
+		case err := <-errorChannel:
+			if err.Errors != nil {
+				for _, e := range err.Errors {
+					t.Errorf("code: %s, message: %s", e.Code, e.Message)
+				}
+			}
+		case log, ok := <-logs:
+			if !ok {
+				break loop
+			}
+			assert.NotNil(t, log.Id)
+
+			assert.NotNil(t, log.Fraud.Id)
+
+			assert.NotEqual(t, "", log.Type)
+
+			assert.NotNil(t, log.Created)
+		}
+	}
+}
+
+func TestPixFraudLogQueryFilters(t *testing.T) {
+
+	starkinfra.User = utils.ExampleProject
+
+	var params = map[string]interface{}{}
+	params["limit"] = 10
+	params["after"] = "2020-04-01"
+	params["before"] = "2030-04-30"
+	params["types"] = []string{"created", "registered"}
+	params["fraudIds"] = []string{"5656565656565656"}
+	params["ids"] = []string{"5656565656565656"}
+
+	logs, errorChannel := PixFraudLog.Query(params, nil)
+loop:
+	for {
+		select {
+		case err := <-errorChannel:
+			if err.Errors != nil {
+				for _, e := range err.Errors {
+					t.Errorf("code: %s, message: %s", e.Code, e.Message)
+				}
+			}
+		case log, ok := <-logs:
+			if !ok {
+				break loop
+			}
+			assert.NotNil(t, log.Id)
+		}
+	}
+}
+
+func TestPixFraudLogPageCursor(t *testing.T) {
+
+	starkinfra.User = utils.ExampleProject
+
+	var params = map[string]interface{}{}
+	params["limit"] = 10
+	params["after"] = "2020-04-01"
+	params["before"] = "2030-04-30"
+	params["types"] = []string{"created", "registered"}
+	params["cursor"] = nil
+
+	logs, cursor, err := PixFraudLog.Page(params, nil)
+	if err.Errors != nil {
+		for _, e := range err.Errors {
+			t.Errorf("code: %s, message: %s", e.Code, e.Message)
+		}
+	}
+
+	for _, log := range logs {
+		assert.NotNil(t, log.Id)
+	}
+	assert.NotNil(t, cursor)
+}